Design: App에서 Main 컴포넌트 분리
